adjusted loops

+++ utils/clean_up_repo.py
@@ -13,8 +13,8 @@ def delete_file(filename):
     if os.path.exists(filename):
         os.remove(filename)
         print(f'{filename}deleted')
-    else:
-        print(f'{filename} has been already deleted')
+    # else:
+        # print(f'{filename} has been already deleted')
 
 def delete_files_in_dir(directory, files):
     "Delete specific files in the directory"
@@ -24,11 +24,16 @@ def delete_files_in_dir(directory, files):
 def delete_files_used_in_example():
     "Delete files used in example from the template"
     for every_dir_list in conf.dir_list:
-        for every_file_list in conf.file_list:
-            delete_files_in_dir(every_dir_list,every_file_list)
+        if os.path.exists(every_file_list):
+            for every_file_list in conf.file_list:
+                delete_files_in_dir(every_dir_list,every_file_list)
+        else:
+            print("All the files related to the sample example from Page Object Model have been removed from tests, conf, page_objects, endpoints folder. For next steps, please refer to the 'edit files' section of this blog post:https://qxf2.com/blog/how-to-start-using-the-qxf2-framework-with-a-new-project/")
+
 
 #----START OF SCRIPT
 if __name__ == "__main__":
     print("Running utility to delete the files")
     delete_files_used_in_example()
-    print("All the files related to the sample example from Page Object Model have been removed from tests, conf, page_objects, endpoints folder. For next steps, please refer to the 'edit files' section of this blog post:https://qxf2.com/blog/how-to-start-using-the-qxf2-framework-with-a-new-project/")
+
+    #print("All the files related to the sample example from Page Object Model have been removed from tests, conf, page_objects, endpoints folder. For next steps, please refer to the 'edit files' section of this blog post:https://qxf2.com/blog/how-to-start-using-the-qxf2-framework-with-a-new-project/")
